LPS-43952 Unnecessary cast

+++ portal-impl/test/integration/com/liferay/portlet/journal/service/persistence/JournalArticleFinderTest.java
@@ -557,7 +557,7 @@ public class JournalArticleFinderTest {
 			article.setReviewDate(sequenceDate);
 			article.setArticleId(sequenceString);
 			article.setTitle(sequenceString);
-			article.setVersion((double)i);
+			article.setVersion(i);
 
 			JournalArticleLocalServiceUtil.updateJournalArticle(article);
 		}
